+++ test/test.py
@@ -451,7 +451,7 @@ class TestEntityFileBP(unittest.TestCase):
 
     def test_add_component_groups(self):
         component_group_data = { "minecraft:damage" : { "value" : 1 } }
-        component_group = self.entity.create_component_group('group:one',component_group_data)
+        component_group = self.entity.add_component_group('group:one',component_group_data)
         self.assertEqual(component_group.id, 'group:one')
 
         saved_bp, saved_rp = save_and_return_packs(bp=self.bp)
@@ -465,19 +465,22 @@ class TestEntityFileBP(unittest.TestCase):
         # Tests before
         self.assertEqual(len(self.entity.components), 29)
         self.assertEqual(len(self.group.components), 4)
+
         # Add to components
         component_data = { "value" : 1 }
-        self.entity.create_component("minecraft:damage", component_data)
+        self.entity.add_component("minecraft:damage", component_data)
         # Add to current component group
-        self.entity.create_component("minecraft:damage", component_data, self.group.id)
-        # Add to new component group
-        self.entity.create_component("minecraft:damage", component_data, "group:one")
+        self.entity.add_component("minecraft:damage", component_data, self.group.id)
+
+        # Add to new component group that doesn't exist
+        with self.assertRaises(AssetNotFoundError):
+            self.entity.add_component("minecraft:damage", component_data, "group:one")
 
         saved_bp, saved_rp = save_and_return_packs(bp=self.bp)
 
         saved_entity = saved_bp.get_entity('minecraft:dolphin')
         self.assertEqual(len(saved_entity.components), 30)
-        self.assertEqual(len(saved_entity.component_groups),8)
+        self.assertEqual(len(saved_entity.component_groups),7)
         saved_group = saved_entity.get_component_group('dolphin_adult')
         self.assertEqual(len(saved_group.components), 5)