+++ reticulator/changelog.md
@@ -57,5 +57,6 @@
   - Getters for all component types
 
 # 0.0.17-beta
- - Fixed EntityFileBP.create_component_group
- - Fixed EntityFileBP.create_component
+ - Replaced EntityFileBP.create_component_group with EntityFileBP.add_component_group
+ - Replaced EntityFileBP.create_component with EntityFileBP.add_component
+ - Replaced ComponentGroup.create_component with ComponentGroup.add_component